requirements.txt
adblockparser==0.7
autocommand==2.2.2
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
cheroot==10.0.1
CherryPy==18.10.0
cherrypy-cors==1.7.0
click==8.1.7
cryptography==3.4.8
diffusers==0.31.0
dnspython==2.7.0
et_xmlfile==2.0.0
ExifRead==2.3.2
filelock==3.16.1
Flask==3.0.3
fsspec==2024.10.0
future==1.0.0
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gunicorn==23.0.0
httpagentparser==1.9.5
httplib2==0.22.0
huggingface-hub==0.26.2
idna==3.8
importlib_metadata==8.5.0
ipaddr==2.2.0
ipwhois==1.1.0
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
Jinja2==3.1.4
joblib==1.4.2
lxml==4.9.4
Mako==1.3.6
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
netaddr==0.10.1
networkx==2.6.3
numpy==2.1.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
phonenumbers==8.13.48
pickle-mixin==1.0.2
pillow==11.0.0
portend==3.2.0
proto-plus==1.24.0
protobuf==5.27.3
publicsuffixlist==0.9.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pygexf==0.2.2
pyOpenSSL==21.0.0
pyparsing==3.1.2
PyPDF2==1.28.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
python-pptx==0.6.23
python-whois==0.7.3
pytz==2024.1
PyYAML==6.0.2
redmail==0.6.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
secure==0.3.0
setuptools==75.2.0
six==1.16.0
soupsieve==2.6
sympy==1.13.1
tempora==5.7.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.1.3
XlsxWriter==3.2.0
zc.lockfile==3.0.post1
zipp==3.21.0
